requirements.txt
accelerate==0.20.3
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
cmake==3.26.4
diffusers==0.20.0
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
fsspec==2023.6.0
gunicorn==20.1.0
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
invisible-watermark==0.2.0
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
psutil==5.9.5
pydantic==2.0.2
pydantic_core==2.1.2
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
#triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
zipp==3.15.0
jinja2
loguru==0.6.0

google-api-python-client==2.43.0
google-api-core #1.31.5
#google-cloud-storage==2.3.0 #not on gae python
google-cloud-storage==2.0.0

google-cloud-ndb==1.11.1
cachetools==4.2.4

python-multipart==0.0.6
nltk==3.8.1
diskcache==5.5.1

protobuf==3.19.5
google-cloud-aiplatform==1.25.0
# openai==0.27.7
# requests==2.28.2
# rollbar==0.16.3
# optimum[openvino]
streamlit
